Moved syncutils up
